Move quantization methods back to  deepseek v3 state_dict_adapter from general protocal, and format files
